requirements.txt
python-telegram-bot==21.1.1
requests==2.31.0
selenium==4.19.0

